+++ CHANGELOG.md
@@ -2,7 +2,26 @@
 All notable changes to this project will be documented in this file.
 `APButton` adheres to [Semantic Versioning](http://semver.org/).
 
-## [3.0.0](https://github.com/APUtils/APButton/releases/tag/3.0.0)
+## [3.1.0](https://github.com/APUtils/APButton/releases/tag/3.1.0)
+Released on 08/07/2017.
+
+#### Added
+- Highlighted for border
+- Disabled for border
+  - Added by [Anton Plebanovich](https://github.com/anton-plebanovich).
+
+#### Fixed
+- Possible leak fix
+  - Added by [Anton Plebanovich](https://github.com/anton-plebanovich).
+
+## [3.0.1](https://github.com/APUtils/APButton/releases/tag/3.0.1)
+Released on 08/03/2017.
+
+#### Fixed
+- Activity animation fix
+  - Added by [Anton Plebanovich](https://github.com/anton-plebanovich).
+
+## [3.0.1](https://github.com/APUtils/APButton/releases/tag/3.0.1)
 Released on 08/02/2017.
 
 #### Added